+++ addon/plugins/article-structure-plugin/commands/recalculate-structure-numbers.ts
@@ -5,13 +5,23 @@ import type { Resource } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/name
 import { ELI } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/constants';
 import { StructureSpec } from '..';
 
-function updateNumber(
-  transaction: Transaction,
-  schema: Schema,
-  { attrs }: PNode,
-  number: string,
-  numberPredicate?: Resource,
-) {
+type UpdateNumberProps = {
+  transaction: Transaction;
+  schema: Schema;
+  node: PNode;
+  number: string;
+  numberPredicate?: Resource;
+};
+
+function updateNumber({
+  transaction,
+  schema,
+  node,
+  number,
+  numberPredicate,
+}: UpdateNumberProps) {
+  const { attrs } = node;
+
   if (isRdfaAttrs(attrs) && attrs.rdfaNodeType === 'resource') {
     const properties = attrs.properties;
     const numberProp = properties.find((prop) =>
@@ -21,6 +31,7 @@ function updateNumber(
       numberProp?.object.termType === 'LiteralNode' && numberProp.object.value;
     const numberNode =
       numberRdfaId && findNodeByRdfaId(transaction.doc, numberRdfaId);
+
     if (numberNode) {
       transaction.replaceWith(
         numberNode.pos + 1,
@@ -38,9 +49,21 @@ export default function recalculateStructureNumbers(
 ) {
   const doc = transaction.doc;
   const indices = new Array<number>(structureSpecs.length).fill(1);
+
   const contexts: (PNode | null)[] = new Array<PNode | null>(
     structureSpecs.length,
   ).fill(null);
+
+  /**
+   * Recording all calls to `updateNumber` in an array to run them all
+   * _after_ we've run all the `setNumber` calls.
+   *
+   * This is necessary because `updateNumber` calls can change positions of the nodes
+   * in the document, and the calls to `setNumber` are dependent on the correct position
+   * before the change.
+   */
+  const updateNumberCalls: UpdateNumberProps[] = [];
+
   doc.descendants((node, pos, parent) => {
     structureSpecs.forEach((spec, i) => {
       if (node.type.name === spec.name) {
@@ -48,20 +71,31 @@ export default function recalculateStructureNumbers(
           indices[i] = 1;
           contexts[i] = parent;
         }
-        if ('convertNumber' in spec.updateNumber) {
-          updateNumber(
+
+        const startNumber = spec.getStartNumber({ pos, transaction });
+
+        if (startNumber) {
+          indices[i] = startNumber;
+        }
+
+        spec.setNumber({ transaction, pos, number: indices[i] });
+
+        if ('convertNumber' in spec.numberConfig) {
+          updateNumberCalls.push({
             transaction,
             schema,
             node,
-            spec.updateNumber.convertNumber(indices[i]),
-            spec.updateNumber.numberPredicate,
-          );
-        } else {
-          spec.updateNumber({ transaction, pos, number: indices[i] });
+            number: spec.numberConfig.convertNumber(indices[i]),
+            numberPredicate: spec.numberConfig.numberPredicate,
+          });
         }
+
         indices[i] += 1;
       }
     });
   });
+
+  updateNumberCalls.forEach(updateNumber);
+
   return transaction;
 }