+++ addon/plugins/variable-plugin/variables/address.ts
@@ -352,6 +352,7 @@ const emberNodeConfig: EmberNodeConfig = {
   uriAttributes: ['variableInstance'],
   draggable: false,
   needsFFKludge: true,
+  selectable: true,
   attrs: {
     ...rdfaAttrSpec({ rdfaAware }),
     value: {