+++ addon/plugins/variable-plugin/variables/codelist.ts
@@ -167,6 +167,7 @@ const emberNodeConfig: EmberNodeConfig = {
   uriAttributes: ['variableInstance'],
   draggable: false,
   needsFFKludge: true,
+  selectable: true,
   attrs: {
     ...rdfaAttrSpec({ rdfaAware }),
     selectionStyle: {