+++ addon/plugins/variable-plugin/variables/text.ts
@@ -130,6 +130,7 @@ const emberNodeConfig: EmberNodeConfig = {
   draggable: false,
   needsFFKludge: true,
   editable: true,
+  selectable: true,
   attrs: {
     ...rdfaAttrSpec({ rdfaAware }),
   },